+++ src/frozen/__init__.py
@@ -1,4 +1,4 @@
-from .freezable import freezableclass, freezablemethod
-from .lockable import lockableclass, lockablemethod
+from .freezable import freezable, freezableclass, freezablemethod
+from .lockable import lockable, lockableclass, lockablemethod
 from .freezable import FrozenError
 from .lockable import LockError, UnlockError, LockedError